+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.1.3 - 2025-09-22
+
+### Added
+
+- Branch-dependent OCV-SOC determination and voltage-based SOC initialisation is now supported for parameter sets using hysteresis and/or multi-material negative electrodes.
+
+### Fixed
+
+- Fixed newly identified PyBaMM [BPX import bug #5193](https://github.com/pybamm-team/PyBaMM/issues/5193) (incorrect porosity import).
+- Fixed incorrect implementation (from v0.1.2) of hysteresis definitions in PyBaMM 25.8.
+
+### Changed
+
+- Minor internal refactors.
+
 ## v0.1.2 - 2025-09-11
 
 ### Fixed